Add checks for potentially nonsemantic class / id names, additional table checks.
Tries to check for classnames and IDs that are potentially
non-semantic.  Presents the actual classname / ID in the notice so you
can quickly see if it's a false alarm.  SOC is touchy, but this is
probably worth at least saying "is this actually what you want?"

Added more table related checks for colgroups, captions and order of
elements.