Try to display error notes for void elements
Void elements (input, img, br) cannot show CSS content.  Instead of
relying on a user inspecting an element to find the associated error
code hidden in a developer tool - try to surface that message with the
next available non-void element.  Still not perfect (if an offending
img tag has no siblings, for example) - but this should help a bit,
especially since the offending image will still be outlined.